--- a/ProtoPromise_Unity/Assets/Plugins/ProtoPromise/Core/InternalShared/ValueCollectionsInternal.cs
+++ b/ProtoPromise_Unity/Assets/Plugins/ProtoPromise/Core/InternalShared/ValueCollectionsInternal.cs
@@ -214,12 +214,6 @@ namespace Proto.Promises
             }
 
             [MethodImpl(InlineOption)]
-            internal T Peek()
-            {
-                return _head;
-            }
-
-            [MethodImpl(InlineOption)]
             internal T Pop()
             {
                 T temp = _head;