--- a/ProtoPromise_Unity/Assets/Plugins/ProtoPromise/Core/Promises/Internal/PromiseFieldsInternal.cs
+++ b/ProtoPromise_Unity/Assets/Plugins/ProtoPromise/Core/Promises/Internal/PromiseFieldsInternal.cs
@@ -245,6 +245,7 @@ namespace Proto.Promises
                     internal SpinLocker _progressCollectionLocker;
                     internal Fixed32 _currentProgress;
                     volatile internal int _progressRetains;
+                    volatile internal bool _handledBranches;
 #endif
                 }