--- a/ProtoPromise_Unity/Assets/Plugins/ProtoPromise/Core/Promises/Internal/PromiseFieldsInternal.cs
+++ b/ProtoPromise_Unity/Assets/Plugins/ProtoPromise/Core/Promises/Internal/PromiseFieldsInternal.cs
@@ -244,8 +244,6 @@ namespace Proto.Promises
 #if PROMISE_PROGRESS
                     internal SpinLocker _progressCollectionLocker;
                     internal Fixed32 _currentProgress;
-                    volatile internal int _progressRetains;
-                    volatile internal bool _handledBranches;
 #endif
                 }