https://github.com/timcassell/ProtoPromise/issues/26#issuecomment-1002876754

Current master:
```
|                  Type |          Method |   N | BaseIsPending |       Mean | Code Size | Allocated | Survived |
|---------------------- |---------------- |---- |-------------- |-----------:|----------:|----------:|---------:|
|          AwaitPending | ProtoPromise_V2 | 100 |             ? | 349.718 μs |  12,107 B |         - |    416 B |
|          AsyncPending | ProtoPromise_V2 | 100 |             ? | 108.983 μs |   7,600 B |         - |    896 B |
|         AsyncResolved | ProtoPromise_V2 | 100 |             ? |  86.302 μs |   5,981 B |         - |    416 B |
|         AwaitResolved | ProtoPromise_V2 | 100 |             ? |   8.802 μs |   2,192 B |      72 B |        - |
|   ContinueWithPending | ProtoPromise_V2 | 100 |         False | 464.729 μs |  18,833 B |         - | 26,984 B |
| ContinueWithFromValue | ProtoPromise_V2 | 100 |         False |  24.830 μs |   8,152 B |         - |        - |
|  ContinueWithResolved | ProtoPromise_V2 | 100 |         False |  31.786 μs |  10,159 B |         - |        - |
|   ContinueWithPending | ProtoPromise_V2 | 100 |          True | 462.657 μs |  15,358 B |         - | 26,832 B |
| ContinueWithFromValue | ProtoPromise_V2 | 100 |          True | 123.617 μs |   9,905 B |         - | 17,160 B |
|  ContinueWithResolved | ProtoPromise_V2 | 100 |          True | 141.513 μs |  10,023 B |         - | 26,760 B |
```
This PR:
```
|                  Type |          Method |   N | BaseIsPending |       Mean | Code Size | Allocated | Survived |
|---------------------- |---------------- |---- |-------------- |-----------:|----------:|----------:|---------:|
|          AwaitPending | ProtoPromise_V2 | 100 |             ? | 363.163 μs |  12,120 B |         - |    416 B |
|          AsyncPending | ProtoPromise_V2 | 100 |             ? |  96.698 μs |   7,531 B |         - |    760 B |
|         AsyncResolved | ProtoPromise_V2 | 100 |             ? |  17.126 μs |   2,149 B |         - |        - |
|         AwaitResolved | ProtoPromise_V2 | 100 |             ? |   8.548 μs |   2,192 B |      72 B |        - |
|   ContinueWithPending | ProtoPromise_V2 | 100 |         False | 477.001 μs |  18,711 B |         - | 24,584 B |
| ContinueWithFromValue | ProtoPromise_V2 | 100 |         False |  25.038 μs |   8,211 B |         - |        - |
|  ContinueWithResolved | ProtoPromise_V2 | 100 |         False |  33.305 μs |  10,139 B |         - |        - |
|   ContinueWithPending | ProtoPromise_V2 | 100 |          True | 451.358 μs |  15,313 B |         - | 24,432 B |
| ContinueWithFromValue | ProtoPromise_V2 | 100 |          True | 120.645 μs |   9,879 B |         - | 17,160 B |
|  ContinueWithResolved | ProtoPromise_V2 | 100 |          True | 136.991 μs |  10,024 B |         - | 24,360 B |
```

Ignore the results of the Async benchmarks, there was a bug that I fixed in this PR where the debug and release optimizations were swapped.

But otherwise, memory usage is a little lower in this PR and code size is slightly smaller. The mean measurements for ContinueWithPending should be taken with a grain of salt, because they are extremely volatile. Code Size should give a more accurate picture.

Memory is also reduced in Race/First/All/Merge promises, but I haven't benchmarked those at all.